+++ changelog.md
@@ -25,7 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Depend on published `svg-dom` crate (`c3d4e5b`)
 - Switch publish flag (`574838f`)
 - Change node labels away from `&'static str` (`15eecf6`)
-- `Scene` must genuinely own or bind to its SVG root (``)
+- `Scene` must genuinely own or bind to its SVG root (`dd421e8`)
+- `NodeId` identifies the graph to which it belongs (``)
 
 ## Added